+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-06-10
+
 ### Added
 
 - New package @react-corn/material-ui-pickers for @material-ui/pickers integration
@@ -24,6 +26,7 @@
 
 ## [1.0.1] - 2020-06-04
 
-[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/paradoxxxzero/react-corn/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/paradoxxxzero/react-corn/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/paradoxxxzero/react-corn/compare/v1.0.0...v1.0.1